Remove old files that should have renamed
The intent of the last edit was to essential rename a bunch of
html files to correspond with their final mailing name, but that
does not appear to have worked.  This commit simply removes the
references to the old files that should have been renamed, losing
any history along the way.