Fix wording to catch 'const' values
Several people pointed out there is a flaw that must be
fixed to avoid accidentally binding to 'const'-qualified
rvalues.  The precedent is already set by 'std::ref' so
simply follow that pattern in the proposed wording.

A second comment was that this facility should be usable
in 'constexpr' contexts, and there is no obvious reason
to not support this.

Motivational text still needs updating.