Clean up formatting and a wording nit
Fix a few broken HTML tags and entities, and reword how we
handle undefined behavior when the invented second iterator
for the algorithms taking only one iterator at the start of
a range cannot reach the invented iterator.  Also reword to
reflect that, for the algorithms in question, it may be either
an input iterator or a forward iterator.