Extensive reformatting and indexing of the document
The r0 document escaped in a much earlier form than I had realized.
The table of contents did not reflect the document, and worse, the
document itself was little more than a bullet-list of topics to
cover.

This commit fixes the majority of the formatting and indexing issues,
and further cleans up a number of wording issues with the requirements
tables.  Still a long way from the finished article.